undo refactor and add getters
