+++ server/src/main/java/com/feeder/server/controller/MultiFeedController.java
@@ -1,11 +1,11 @@
 package com.feeder.server.controller;
 
-import com.feeder.server.model.data.DemoData;
-import com.feeder.server.model.data.GenericData;
-import com.feeder.server.model.data.GithubData;
-import com.feeder.server.model.data.HackerNewsData;
-import com.feeder.server.model.data.RedditData;
-import com.feeder.server.model.data.TwitterData;
+import com.feeder.server.model.DemoData;
+import com.feeder.server.model.GenericData;
+import com.feeder.server.model.GithubData;
+import com.feeder.server.model.HackerNewsData;
+import com.feeder.server.model.RedditData;
+import com.feeder.server.model.TwitterData;
 import com.feeder.server.provider.FeedProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
