+++ server/src/main/java/com/feeder/server/model/CovidNineteenData.java
@@ -1,10 +1,10 @@
-package com.feeder.server.model.data;
+package com.feeder.server.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-import com.feeder.server.model.data.AutoValue_CovidNineteenData;
+import com.feeder.server.model.AutoValue_CovidNineteenData;
 import com.google.auto.value.AutoValue;
 
 /** A CovidNineteenData type represents properties of the web scraped covid19 data */
