+++ server/src/main/java/com/feeder/server/model/DemoData.java
@@ -1,11 +1,11 @@
-package com.feeder.server.model.data;
+package com.feeder.server.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-import com.feeder.server.model.data.AutoValue_DemoData;
+import com.feeder.server.model.AutoValue_CovidNineteenData;
 import com.google.auto.value.AutoValue;
 import java.util.Optional;
