+++ server/src/main/java/com/feeder/server/provider/FeedProvider.java
@@ -1,6 +1,6 @@
 package com.feeder.server.provider;
 
-import com.feeder.server.model.data.GenericData;
+import com.feeder.server.model.GenericData;
 import reactor.core.publisher.Flux;
 
 /**
