+++ server/src/test/java/com/feeder/server/provider/hackernews/HackerNewsFeedProviderTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.feeder.server.model.data.HackerNewsData;
+import com.feeder.server.model.HackerNewsData;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
